file structure amended to allow for testing and cors
